Adds users app
