+++ auto_selecto/concesionariaApp/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from . import models
+
 
 # Register your models here.
