+++ auto_selecto/concesionariaApp/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path, include
+from rest_framework import routers
+
+router = routers.DefaultRouter
+#router.register()
+urlpatterns = [
+    path("api/v1 ")
+]
