+++ auto_selecto/users/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Client, User
+# Register your models here.
+admin.site.register(User)
+admin.site.register(Client)
