+++ auto_selecto/users/serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from . import models
+
+class ClientSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.Client
+        fields = '__all__'
