+++ auto_selecto/users/urls.py
@@ -1,0 +1,4 @@
+from django.urls import path
+from . import views
+
+#here is all code to manage users
