+++ auto_selecto/users/views.py
@@ -3,4 +3,5 @@ from rest_framework import viewsets
 from ..users.serializer import ClientSerializer
 
 # Create your views here.
-
+class ClientView(viewsets.ModelViewSet):
+    ClientSerializer
